Added error checking, Not sure what the mark items issue is, will turn this new verion in then check tomorrow with Larose

+++ todo_list/main.py
@@ -28,12 +28,15 @@ def main():
     list_loc1 = "todo_list/todo_checks.txt" # Stores the info that says your task is complete
 
     create_list(list_loc,list_loc1) # Creates list if no list is there
-
+    error = ''
     while True:
         print('\033c') # Clears terminal at the beggining
         to_do_list, to_do_check, chosen = read_file(list_loc,list_loc1) # Reads the file and returns the list, checks, and if your task is complete "X" " "
         
         display(to_do_list,to_do_check) # Displays to-do list
+        if error:
+            print(error) # Prints an error message for the user
+        error = ''
 
         action = inquirer.select(
             message="What would you like to do?",
@@ -59,23 +62,29 @@ def main():
                 ).execute())
             
         elif action == 'mark':
-            mark_item(to_do_list,to_do_check,
-                inquirer.checkbox(
-                    message="What would you like to mark off?",
-                    # Gives you each of the tasks, then checks if task is checked and uses check_to_bool to do so, LIST COMPREHENSION IS COOL
-                    choices=[Choice(x.strip(),enabled=check_to_bool(chosen[to_do_list.index(x)-1])) for x in to_do_list[1:]],
-                    instruction="Press [Space] to select, and [Enter] to finalize selction",
-                ).execute())
-            
+            if len(to_do_list) > 1:
+                mark_item(to_do_list,to_do_check,
+                    inquirer.checkbox(
+                        message="What would you like to mark off?",
+                        # Gives you each of the tasks, then checks if task is checked and uses check_to_bool to do so, LIST COMPREHENSION IS COOL
+                        choices=[Choice(x.strip(),enabled=check_to_bool(chosen[to_do_list.index(x)-1])) for x in to_do_list[1:]],
+                        instruction="Press [Space] to select, and [Enter] to finalize selction",
+                    ).execute())
+            else:
+                error = 'Please add an item to the to-do list before marking items as complete!\n'
+
         elif action == 'delete':
-            delete_item(to_do_list, to_do_check,
-                        inquirer.fuzzy(
-                            message="Choose an item to delete",
-                            # Same thing as in mark, but it doesn't have a check box
-                            choices=[Choice(value=x, name=x.strip()) for x in to_do_list[1:]],
-                            validate=lambda result:result in to_do_list and result != "Your To-Do List:",
-                        ).execute())
-            
+            if len(to_do_list) > 1:
+                delete_item(to_do_list, to_do_check,
+                            inquirer.fuzzy(
+                                message="Choose an item to delete",
+                                # Same thing as in mark, but it doesn't have a check box
+                                choices=[Choice(value=x, name=x.strip()) for x in to_do_list[1:]],
+                                validate=lambda result:result in to_do_list and result != "Your To-Do List:",
+                            ).execute())
+            else:
+                error = 'Please add an item to the to-do list before attempting to delete an item!\n'
+
         elif action == 'restart': # Resets the list, deletes all items
             to_do_list = ["Your To-Do List:"]
             to_do_check = ["Your Check List:"]
